requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.0
asyncio-mqtt==0.16.1
cors==1.0.1
